requirements.txt
attrs==17.3.0
email-validator==1.0.3
Inject==3.3.2
pycodestyle==2.3.1
pytest==3.3.1
validator.py==1.2.5
